成功生成除了highpriority之外的所有东西

+++ com.rm2pt.generator.cloudedgecollaboration/src/com/rm2pt/generator/cloudedgecollaboration/generator/lyz/ServiceParser.java
@@ -107,7 +107,7 @@ public class ServiceParser {
         if (type instanceof BasicType) {
             switch (((BasicType) type).getTypeEnum().name()) {
                 case "INTEGER":
-                    return "int32";
+                    return "int64";
                 case "TIME":
                     return "time.Time";
                 case "REAL":
